Adds alias to the webpack for paths
Previously I was specifying the full paths of files I was requiring. This
commit adds alias to the webpack which allows to specify the file path only in
one place. This will make changing the file directory much easier in the future
as the path will have to be changed only in one place.